authrization added

+++ backend/routes/post.js
@@ -27,10 +27,13 @@ const storage = multer.diskStorage( {
 });
 
 router.post("",checkAuth, multer({ storage : storage}).single('image'),(req,res,next)=>{
+    console.log( req.userData);
+
     const url = req.protocol + "://"+req.get('host');
     const post =  new Post( {title : req.body.title
         , content : req.body.content
         , imagepath : url+'/images/' + req.file.filename
+        , creator : req.userData.userid
     });
     console.log(post);
     post.save().then((result)=>{
@@ -39,7 +42,7 @@ router.post("",checkAuth, multer({ storage : storage}).single('image'),(req,res,
             id : result._id,
             title : result.title,
             content : result.content,
-            imagepath : result.imagepath 
+            imagepath : result.imagepath
         } }); 
     });
 });
@@ -93,10 +96,14 @@ router.get("/:id",(req,res,next )=>{
 
 router.delete('/:id',checkAuth, (req , res, next) => {
     console.log('deleting  id : ',req.params.id);
-    Post.deleteOne( { _id : req.params.id } ).then( result => {
-            console.log("Delete Result "+ result);    
+    Post.deleteOne( { _id : req.params.id ,  creator : req.userData.userid } ).then( result => {
+        if(result.n > 0){
+            console.log(" Update Result "+ result);    
             res.status(200).json( { message : 'Post deleted' } );
-        });
+        }else{ 
+            return res.status(401).json( { message : 'Your are not autherized to delete content' });
+        } 
+    });
 } );
 
 router.put('/:id',checkAuth, multer({ storage : storage}).single('image'), (req , res, next) => {
@@ -105,11 +112,15 @@ router.put('/:id',checkAuth, multer({ storage : storage}).single('image'), (req 
         const url = req.protocol + "://"+req.get('host');
         imagepath = url+'/images/' + req.file.filename;
     }
-    const post  = { _id : req.body.id ,title : req.body.title , content : req.body.content , imagepath : imagepath};
+    const post  = { _id : req.body.id ,title : req.body.title , content : req.body.content , imagepath : imagepath, creator : req.userData.userid};
     console.log('edit id : ',req.params.id);
-    Post.updateOne( { _id : req.params.id}, post ).then( result => {
-            console.log(" Update Result "+ result);    
-            res.status(200).json( { message : 'Post updated' } );
+    Post.updateOne( { _id : req.params.id , creator : req.userData.userid }, post ).then( result => {
+            if(result.nModified > 0){
+                console.log(" Update Result "+ result);    
+                return res.status(200).json( { message : 'Post updated' });
+            }else{ 
+                return res.status(401).json( { message : 'Your are not autherized to update content' });
+            }
         });
 } );
